[ffx] Add In-use handle value type to Playground
The Rust bindings for Zircon handles, and the value type we get from
FIDL codec, both treat handles as single-owner. In the shell environment
this doesn't quite work. To get around this, we add an in-use handle
value type, which is sort of a ref-counted container with a handle
inside of it. We can duplicate this value as much as we want, all the
copies reference the same handle, and any one of them can steal the
handle if we need to send it through a channel for example, with the
others inducing graceful failures if they're operated on further.

Change-Id: Ib02b26e5a42f3ba3a135141d968c050d05c31a6f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987295
Reviewed-by: Andrew Davies <awdavies@google.com>