[performance][docs] Update docs for creating metric name expectation files
Update these docs to work for Python-based tests.

Drop the description of doing this for Dart-based tests.  We generally
won't be adding new Dart-based tests, so we don't need to complicate
the docs with covering that.

These docs are generally complicated by the fact that GN will complain
if you specify (in GN) an expectation file that doesn't exist yet.  An
alternative here might be to add a GN argument for telling the tests
to create these expectation files.

Bug: b/307980390
Change-Id: I0ee5fc459a3f1058fa99f933a476b44fb81d4e99
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/992753
Commit-Queue: Mark Seaborn <mseaborn@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>