[pbms]Improving product bundle error messages
This improves the error messages when loading a product bundle.

The functionality is also made more friendly to command line specification of the product bundle path.

A Product Bundle's path can be specified by a relative path. If this
path relative to the current directory does not exist, the relative
path resolved relative to the build directory, if one exists (either
in-tree or if configured via ConfigDomain context).

This also removes the use of the SDK configuration object when loading.

Bug: b/288891258
Bug: b/325271553
Change-Id: I6d3f2ce344a7abe8f2f19c6c35381331b221437c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/992655
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>