[kernel][arm64] Use <lib/arch/arm64/exception-asm.h> fully
This finishes the transition to the assembly macro scheme with
.vbar_table and .vbar_function already used in the phys code.
The instruction sequence and layout of the vector code is not
changed, only the assembly source uses less boilerplate and some
symbol and section names are changed in a uniform scheme.  This
paves the way for moving the code around and further macroifying
things more easily, including the possibility of multiple
variants of the vector table with minimal source duplication.

The only actual change in the disassembly is the "vector number"
printed on the panic path for invalid exceptions, which simply
uses the prescribed offset from the vector base address to
indicate which hardware vector it is, rather than the previous
idiosyncratic set of numbers.

Bug: 323463901
Change-Id: I6dcb1fbc88ded32171f54fe928830ef11c3ff52e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/985732
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Marco Vanotti <mvanotti@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>