[rbe] Attach unique invocation_id to each action
Then we won't need a reproxy log post-processing step to attach
a build id to each LogRecord.

fuchsia-reproxy-wrap.sh: `build_uuid` is now available even when
auto-uploading metrics+logs is disabled.
Set rewrapper `invocation_id` with a "user:" prefix, so we can know
that the origin was a developer build, as opposed to an infra build.

Bug: b/325521256
Change-Id: Id2d893075255343d83d1e6d1fc81552f367f87f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/992517
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Fang <fangism@google.com>