[debugger] Display "PID"/"TID" on Linux.
When the debugged process is Linux, use the platform names "PID" and
"TID" for process and thread IDs instead of the Fuchsia "KOID" term.

Some tests results changed from Fuchsia to Posix style because the test
didn't set a platform and the default behavior is Posix-like names. I
updated the test expectations rather than setting the platform because
this gives some coverage to these cases (the Fuchsia formatting is used
more often in general).

Change-Id: I619d40455628facd212e0bb0d2ea02a946aa7332
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/985406
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>