[dl] Pull in testing modules from libdl
This CL pulls in the GN logic for libdl to use libld testing modules and
switches so that dl-tests can access the test files.

System tests can now retrieve a file from the test package.

The `kHasDlOpen` testing flag is removed; support for file retrieval in
libdl will be added in a forthcoming CL.

Bug: 324650368
Change-Id: Ib3812d9620247ca4bf66fa1344c406023b82a2d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/991257
Commit-Queue: Caslyn Tonelli <caslyn@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>