[amlogic-display][dsi] Inject PanelConfig to DsiHost.
This adds the PanelConfig const pointer to DsiHost ctor so that the
PanelConfig will be injected and the reference will be stored
in DsiHost ctor.

The PanelConfig will be converted to a display_setting_t and sent
to the DSI Host controller driver to configure the DSI transmitter.

Test: no behavior change expected.
Test: core.nelson booted.
Bug: 318479388
Change-Id: Iadbc1b30b4b460a965743711f6daf0ef86edc369
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/969344
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>
Reviewed-by: Victor Costan <costan@google.com>