[component_manager] Make Router async
Update Router::route from a synchronous function that uses
a completer to an async function. This cleans up the places
in the component manager codebase where the component's
task group had to be use to create an async context.

The main reason the change to async is better is because
it is a useful pattern to write a Router that calls
another router to get a capability *and then* operate
on that capability. This is much easier in an async
context.

This code is a bit awkward because Rust doesn't support
async llambdas. A good clean up here would be to replace
some of these llambdas with structs that impl Routable.
Then Router could hold a Box<dyn Routable>.

Bug: b/325096862
Change-Id: Ie49726a37206b0fc6e1866cdb464a36032a24ac6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/992415
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>