[compat] Lower log severity
This log can occur when stop was triggered before the init hook, but the
driver tries to complete it's init hook once triggered. This sequence is
valid and the driver did nothing wrong, so this shouldn't be logged as
an error.

Change-Id: I121b50664094a85a400f9f77a51a071faacebefd
Fixed: 318983912
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/994664
Reviewed-by: Sarah Chan <spqchan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>