[tests][golden_files] Allowing overriding of test label
This adds a parameter to golden_files template named "visible_label"

Some usages of golden_files do not have global visibility, as a result
the error message printing the label to make sure is included in the
build graph could be not visible.

The visible_label parameter is used to specify a visible label for this
message. If it is not given, the label of the golden_file test is used
as normal.

Bug: b/326258617
Change-Id: I140efcab237e47b3257b350f2781b6d620db61a9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/994961
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>