[filter.deprecated] Add FIDL for probing filter
Add a method to the Filter protocol to check whether there is a server
attached to the other end. This will be used by netcfg to check whether
the optional capability it has is actually being served before using it
to create clients. If the optional capability is present, then netcfg
will use the deprecated protocol instead of the new filter protocol.
The bug to track the usage of that API is https://fxbug.dev/323950209

Bug: 323950209
Fixed: 323949472
Change-Id: I124235810a57773465ef5da5eac29effe2715e83
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987244
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
API-Review: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>