[kernel][phys] Add ArchPatchInfo type and plumbing
This adds the ArchPatchInfo type in physboot.  This will collect
any state that physboot should calculate once and then might be
needed be ArchPatchCode and/or ArchHandoff.  It's separate from
ArchPhysInfo because this is for data specific to physboot's
patching of the kernel proper and any arch-specific handoff logic
that may interrelate with that data, while ArchPhysInfo is for
things that the low-level startup of any phys executable will do.

Bug: 322202704
Change-Id: I6d196ff6cec56d9b565065d57ae1879398659e0d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995832
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>