[netdevice] Keep test events alive during calls
The network device library has a few events that are used for testing
purposes. These events are represented by fit::functions. A recent
change to fix a flake started using one of these events to ensure that
tests were correctly waiting for the network device start sequence to
finish. The fix included replacing an event handler during the
invocation of the same event handler. This introduced a problem where
because of this the old event handler, currently executing, was
destroyed before it completed. The event handler could then attempt to
access destroyed state. In some instances this lead to a captured this
pointer being null, even if the object it pointed to was still alive.

Fix this by wrapping each call to these event handlers in a function
that first shares the event handler and maintains a local reference for
the duration of the invocation. This way the event handler can safely be
replaced without being destroyed until the invocation of the event
handler has finished.

Test: Verified that this type of behavior does not destruct captures.
Bug: b/323601448
Multiply: network-device-lib-test
Change-Id: I63352e0d1ea4c680e40f3c55dd102eb6da222360
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995507
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Bjoern Johansson <bjoernj@google.com>