[archivist] Fix issue that caused NS3 data to be unavailable in pipelines
Without this fix we would remove the NS3 static selectors when we
handled the diagnostics dir closed event. The diagnostics dir and the
InspectSink handle were treated as coming from different components,
even if they were the same, due to the difference in the URL with which
they were launched.

An alternative fix would be to ensure that the URL hardcoded in the
Archivist is the same as the URL in netstack CMLs. However, I think this
change is correct regardless, as we typically care about the moniker
only, not the URL with which the component was launched at that
location. That's just metadata.

Multiply: archivist-integration-tests
Multiply: archivist-unittests
Fixed: 326510415
Change-Id: Ie52b4a058e7e9cbcf8d04f3ee515b30329197536
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995813
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>