[logs] Add support for transactional symbolizer
This CL adds support for the transactional symbolizer
protocol to the log_formatter's definition of what a symbolizer is.
Future patches will implement the log_formatter changes
for handling transactions and the transactional symbolizer itself.
This merely defines the interface to the transactional symbolizer.
The new method exists to facilitate a soft transition to the
transactional symbolizer so it can be enabled or disabled
at runtime with a flag. If the symbolzer is not transactional,
all logging works the same as it has before. For log_listener,
symbolize is a no-op so it can safely claim to support the transactional
path. The old symbolizer won't support transactions so returns
false to indicate to log_formatter that it is invalid to
attempt transactions on it.

Bug: b/299980894
Change-Id: If1109450c263d7658d2dd60da9afd13c143d914c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/996363
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>