[kernel] Count wasted preemption events
Add two new kcounters.  One is incremented whenever the scheduler sets
the preemption timer to a point that's prior to the time latched when
the CPU last entered the scheduler.  The other is incremented when the
scheduler dequeues a deadline task with a finish_time that's prior to
the eligible time (now).  See https://fxbug.dev/42182770 for more
context.

Bug: 42182770
Change-Id: Ib8adb697fb8ee0a6f73c1c29957b9434b28be0f2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995494
Fuchsia-Auto-Submit: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Nick Maniscalco <maniscalco@google.com>