[starnix][netlink] Parse and log SOCKDIAG messages
Previously, there was an issue where by reading bytes from the
InputBuffer, the caller would expect that the `write` completed
successfully even if an error was returned. We can use `peek` to obtain
the bytes and parse the message without removing the bytes from the
buffer.

This allows us to still see which SOCKDIAG messages are being received.

Bug: 323590076
Change-Id: I4b578f2576659742ebb1d9485095612a56c3f3d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995503
Reviewed-by: Dylan Swiggett <swiggett@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>