[bt][core] move LegacyLowEnergyScanner impl to base class
This reverts commit ad3a94a29a8c689a0328dd32bc151dd5521c132b.

This change moves much of the LegacyLowEnergyScanner implementation up
to its base class, LowEnergyScanner. This change also updates tests to
be generic type parameterized tests with LegacyLowEnergyScanner being
the only current tested implementation. This is in preparation of adding
a new ExtendedLowEnergyScanner subclass to LowEnergyScanner.

This change also includes a fix to a LowEnergyScanner crash. If a scan
response timeout occurred before we received a scan response
(e.g. scannable peer which responded slowly with a scan response), we
would have destroyed all context regarding the peer while still trying
to handle the scan response. This handling had a bug where we wouldn't
check whether the peer context was still in our memory or had already
been destructed. When already destructed, we would access now invalid
memory and crash.

As a part of this change, we make a large refactor, moving some packet
builder methods from FakeController to FakePeer. FakePeer by default
sends its advertising reports immediately to facilitate
scanning. However, tests can now also request FakePeer to not do so and
retain control of sending the advertising reports themselves. This
allowed us to write a new test to test the LowEnergyScanner crashed
mentioned in the paragraph above.

We also clean up some of the code on batching advertising data with scan
responses in advertising reports. The majority of the tests followed the
simple path of not batching advertising data with scan responses in the
advertising reports. For the one case where we wanted to batch the data,
we introduce a new test so that we can continue to test such a situation
while also simplifying the code.

Bug: b/323098126
Test: fx test src/connectivity/bluetooth/core/bt-host; added tests
Change-Id: I88881fa093691b019f6605a019c22263d295057f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986428
Reviewed-by: Ben Lawson <benlawson@google.com>
Commit-Queue: Faraaz Sareshwala <fsareshwala@google.com>